pyproject.toml
[tool.poetry]
name = "funes"
version = "0.1.0"
description = "A python package for big time series data information extraction"
authors = ["binjian.xin <binjian.xin@newrizon.com>"]
maintainers = ["binjian.xin <binjian.xin@newrizon.com>"]
readme = "README.md"
license = "LICENSE"
repository = "ssh://git@gitlab.newrizon.work:2222/its/ai/fune:w.git"
keywords = ["Deep Learning", "Generative Model", "Battery", "Outlier Detection", "policy"]
classifiers = [
    "Frameworks :: Pytest",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "3.10.8"
torch = ">=1.11.0"
tensorboard = ">=2.8.0"
python-json-logger = ">=2.0.2"
numpy = ">=1.22.3"
scipy = ">=1.7.0"
pandas = ">=1.4.2"
jupyterlab = ">=3.3.4"
seaborn = ">=0.11.2"
scikit-learn = ">=1.0.2"
matplotlib = ">=3.5.1"
tqdm = ">=4.64"
umap-learn = ">=0.5.3"
pytest = ">=7.1"
elasticsearch = ">=7.1.2"
openpyxl = ">=3.0.9"
absl-py = ">=1.0.0"
ml_collections = ">=0.1.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"